Remove `mmr` from `collection` crate
