--- lib/collection/src/collection/mmr/mod.rs
@@ -1,4 +1,0 @@
-mod lazy_matrix;
-mod maximal_marginal_relevance;
-
-pub use maximal_marginal_relevance::mmr_from_points_with_vector;
